services/deepgram.py
import os
from dotenv import load_dotenv

load_dotenv()

DEEPGRAM_API_KEY = os.getenv("DEEPGRAM_API_KEY")
DEEPGRAM_WS_URL = "wss://api.deepgram.com/v1/listen?encoding=mulaw&sample_rate=8000"

def get_deepgram_headers():
    return [("Authorization", f"Token {DEEPGRAM_API_KEY}")]
