requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.15
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
decorator==5.2.1
distro==1.9.0
elevenlabs==1.55.0
executing==2.2.0
fastapi==0.115.12
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jedi==0.19.2
jiter==0.9.0
matplotlib-inline==0.1.7
multidict==6.3.0
openai==1.74.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.50
propcache==0.3.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.1
pydantic_core==2.33.0
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.1
tqdm==4.67.1
traitlets==5.14.3
twilio==9.5.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.18.3
